+++ CHANGELOG.md
@@ -2,6 +2,28 @@
 
 pandoc-plot uses [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
+Release 1.0.2.0
+---------------
+
+* Added the ability to set command-line arguments for interpreters via configuration. For example, if you want to run the Matplotlib toolkit with all warnings shown:
+
+```yaml
+# The possible parameters for the Matplotlib toolkit
+matplotlib:
+  executable: python
+  command_line_arguments: -Wa
+```
+
+Or if you want `julia` to use more than one thread:
+
+```yaml
+# The possible parameters for the Matplotlib toolkit
+plotsjl:
+  executable: julia
+  command_line_arguments: --threads auto
+```
+
+
 Release 1.0.1.0
 ---------------